requirements.txt
fastapi
uvicorn[standard]
websockets
kafka-python
redis
streamlit
pandas
matplotlib
